backend/requirements.txt
# MALIN Platform https://malin.cahiersfantastiques.fr/
# Copyright 2024-2025 Vincent Jacques <vincent@vincent-jacques.net>

# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU Affero General Public License as published
# by the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.

# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU Affero General Public License for more details.

# You should have received a copy of the GNU Affero General Public License
# along with this program.  If not, see <https://www.gnu.org/licenses/>.

alembic  # Database migrations
argon2-cffi  # Password hasher
black  # Code formatter, used by Alembic
boto3  # AWS, used e.g. for database backups to S3
click  # Command-line parsing
fastapi  # API framework
httpx  # For unit testing: https://fastapi.tiangolo.com/tutorial/testing/
mypy  # Static type checker
psycopg2-binary  # For PostgreSQL
pyhumps  # Convert between cases
pyjwt  # For authentication tokens
PyPDF2  # Only for test fixtures for now
python-multipart  # OAuth2 authentication uses multipart
sqids  # Opaque IDs
sqlalchemy  # Database ORM
sqlalchemy-data-model-visualizer  # Draw database schema
sqlalchemy-utils  # Database utilities
uvicorn[standard]  # ASGI server
